Awaiting blog to get returned before deletion

+++ src/blogs/hooks/useDeleteBlog.js
@@ -1,19 +1,26 @@
 import {getBlogById, deleteBlog} from '../services/blogService'
 
 
+
+// Delete blog hook
 const useDeleteBlog = async (id, token) => {
-    const confirmDeletion = window.confirm("Continue with deleting blog?")
-    if(!confirmDeletion) return;
+    const confirmDeletion = window.confirm("Continue with deleting blog?");
+    if (!confirmDeletion) return;
+
     try {
-        const blog = getBlogById(id)
-        await deleteBlog(blog._id, token)
-        
-        return true
+        const blog = await getBlogById(id);  
+
+        if (!blog) {
+            alert("Blog not found");
+            return false;
+        }
+
+        await deleteBlog(blog._id, token);
+        return true;
     } catch (error) {
-        console.log(error.message)
-        return false
+        console.log(error.message);
+        return false;
     }
-   
-}
- 
+};
+
 export default useDeleteBlog;
